requirements.txt
Flask==1.1.1
gunicorn==20.0.4
line-bot-sdk==2.2.1
APScheduler==3.6.1
psycopg2==2.9.3
